✨ Add JKube custom OCP deploy strategy
