+++ fuse-products/src/main/java/software/tnb/product/deploystrategy/OpenshiftDeployStrategyType.java
@@ -3,5 +3,6 @@ package software.tnb.product.deploystrategy;
 public enum OpenshiftDeployStrategyType {
     JKUBE,
     DEVFILE,
-    BINARY
+    BINARY,
+    CUSTOM
 }
